fix accidental reduction of a real(8) to real(4) (there were some really weird effects due to this little bug....)

+++ fem/src/MagnetoDynamicsUtils.F90
@@ -83,7 +83,7 @@
     TCoefIm = GetElectricConductivityTensor(Element,n,'im',CoilBody,CoilType)
     DO i=1,3
        DO j=1,3
-          Tcoef( i,j,1:n ) = CMPLX( REAL(TcoefRe( i,j,1:n )), TCoefIm( i,j,1:n ), KIND=dp)
+          Tcoef( i,j,1:n ) = CMPLX( TcoefRe( i,j,1:n ), TCoefIm( i,j,1:n ), KIND=dp)
        END DO
     END DO
